Refactor AboutPage to utilize new shared components: HeroSection, MeetTherapist, MyStudio, ContactInfo, and CTASection, enhancing modularity and maintainability.
