+++ app/components/shared/HeroSection.tsx
